--- .env-template
@@ -1,3 +1,0 @@
-# Rename this file to .env and fill in the values for the following variables. It is only required for the web retriver as of now
-GOOGLE_CSE_ID=""
-GOOGLE_API_KEY=""
