+++ builder.py
@@ -8,9 +8,9 @@ from langchain.text_splitter import RecursiveCharacterTextSplitter
 from typing import Optional, Iterator, List, Dict
 from chromadb.config import Settings
 import argparse
-from vectorstore.chroma import ChromaStore
-from vectorstore.faiss import FAISSStore
-from config import (
+from neogpt.vectorstore.chroma import ChromaStore
+from neogpt.vectorstore.faiss import FAISSStore
+from neogpt.config import (
     SOURCE_DIR,
     INGEST_THREADS,
     DEVICE_TYPE,
