+++ neogpt/documents/builder.url
@@ -1,0 +1,1 @@
+https://www.youtube.com/watch?v=BPknz-hCnec
