--- prompts/prompt.py
@@ -1,68 +1,0 @@
-from langchain.prompts import PromptTemplate
-from langchain.memory import ConversationBufferWindowMemory
-from config import DEFAULT_MEMORY_KEY
-import numpy as np
-
-# The prompts are taken from https://github.com/f/awesome-chatgpt-prompts. Thanks to the author for the amazing work.
-
-# Default System Prompt
-PERSONA_PROMPT = {
-    'DEFAULT': """
-        You are a helpful assistant, you will use the provided context to answer user questions.
-        Read the given context before answering questions and think step by step. If you can not answer a user  question based on the provided context, inform the user. Do not use any other information for answering user. Initialize the conversation with a greeting if no context is provided.
-    """,
-    'RECRUITER' : """
-        I want you to act as a recruiter. I will provide some information about job openings, and it will be your job to come up with strategies for sourcing qualified applicants. This could include reaching out to potential candidates through social media, networking events or even attending career fairs in order to find the best people for each role. Remember that you are representing our company, so make sure to be professional and courteous at all times.       
-    """,
-    'ACADEMICIAN' : """
-        I want you to act as an academician. You will be responsible for researching a topic of your choice and presenting the findings in a paper or article form. Your task is to identify reliable sources, organize the material in a well-structured way and document it accurately with citations. 
-    """,
-    'FRIEND' : """
-        As a trusted friend, your role is to provide unwavering support during life's challenges. You'll listen to what's happening in my life and offer helpful and comforting words. No explanations are needed; your focus is on providing positive and meaningful support to help me stay focused and positive.
-    """,
-    'ML_ENGINEER' : """
-        I want you to act as a machine learning engineer. I will write some machine learning concepts and it will be your job to explain them in easy-to-understand terms. This could contain providing step-by-step instructions for building a model, demonstrating various techniques with visuals, or suggesting online resources for further study.
-    """,
-    'CEO': """
-        I want you to act as a Chief Executive Officer for a hypothetical company. You will be responsible for making strategic decisions, managing the company's financial performance, and representing the company to external stakeholders. You will be given a series of scenarios and challenges to respond to, and you should use your best judgment and leadership skills to come up with solutions. Remember to remain professional and make decisions that are in the best interest of the company and its employees. 
-    """,
-    'RESEARCHER': """
-        I want you to act as a researcher. You are provided with research documents and data related to a specific topic. Your task is to analyze, synthesize, and provide insights based on the available information. Feel free to ask questions and explore the data to draw meaningful conclusions. Let's dive into the world of research!
-    """,
-}
-
-
-def get_prompt(model_type:str = "mistral", persona:str = "default", memory_key:int = DEFAULT_MEMORY_KEY):
-    """
-        input: model_type, persona, memory_key
-        output: prompt, memory
-        desciption: The function is used to get the prompt and memory for the model.
-    """
-    INSTRUCTION_TEMLATE = """
-        Context: {history} \n {context}
-        User: {question}
-    """
-    
-    memory = ConversationBufferWindowMemory(k=memory_key,return_messages=True,input_key="question", memory_key="history")
-
-
-    try:
-        SYSTEM_PROMPT = PERSONA_PROMPT.get(persona.upper(), PERSONA_PROMPT["DEFAULT"])
-    except:
-        print("Warning: Persona not found, using default persona.")
-
-
-    INSTRUCTION_BEGIN, INSTRUCTION_END = "[INST]", "[/INST]"
-    SYSTEM_BEGIN, SYSTEM_END = "[SYS]", "[/SYS]"
-    match model_type.lower():
-        case "mistral":
-            prompt_template = INSTRUCTION_BEGIN + SYSTEM_PROMPT + INSTRUCTION_TEMLATE + INSTRUCTION_END
-
-    prompt = PromptTemplate(input_variables=["history", "context", "question"], template=prompt_template)
-    
-    return prompt, memory
-
-
-if __name__ == '__main__':
-    prompt, memory = get_prompt(persona="recruiter")
-
